fastapi_app copy/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
pillow>=10.0.0
pillow-heif>=0.13.0
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.10.0
numpy>=1.24.0,<2.0.0
transformers>=4.30.0
opencv-python>=4.8.0
scikit-image>=0.21.0
fal-client>=0.4.1
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

